+++ zihuan_agent/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod brain;
-pub mod dream_agent;
 pub mod emotion;
 pub mod session_state;
 pub mod utils;
